expanding support for shared memory - first rewrite
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@10219 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/camd/alta.cpp
@@ -56,7 +56,7 @@ class Alta:public Camera
 
 		virtual int setValue (rts2core::Value * old_value, rts2core::Value * new_value);
 
-		virtual long suggestBufferSize ()
+		virtual size_t suggestBufferSize ()
 		{
 			return Camera::suggestBufferSize () + 4;
 		}
